Refactor: properties added
