+++ ApiDatastructures/Credentials/Create/ResponseBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Credentials.Create;
 
 public class ResponseBody
 {
-  public string CredentialUuid;
+  public string CredentialUuid { get; set; }
 }
