+++ ApiDatastructures/Credentials/Delete/RequestBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Credentials.Delete;
 
 public class RequestBody
 {
-  public string CredentialUuid;
+  public string CredentialUuid { get; set; }
 }
