+++ ApiDatastructures/Credentials/Get/RequestBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Credentials.Get;
 
 public class RequestBody
 {
-  public string uuid;
+  public string Uuid { get; set; }
 }
