+++ ApiDatastructures/Credentials/Get/ResponseBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Credentials.Get;
 
 public class ResponseBody
 {
-  public Credential Credential;
+  public Credential Credential { get; set; }
 }
