+++ ApiDatastructures/Credentials/GetAll/ResponseBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Credentials.GetAll;
 
 public class ResponseBody
 {
-  public List<Credential> Credential;
+  public List<Credential> Credential { get; set; }
 }
