+++ ApiDatastructures/Credentials/Update/RequestBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Credentials.Update;
 
 public class RequestBody
 {
-  public CredentialRequest Credential;
+  public CredentialRequest Credential { get; set; }
 }
