+++ ApiDatastructures/Credentials/Update/ResponseBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Credentials.Update;
 
 public class ResponseBody
 {
-  public CredentialResponse Credential;
+  public CredentialResponse Credential { get; set; }
 }
