+++ ApiDatastructures/Folders/Create/RequestBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Folder.Create;
 
 public class RequestBody
 {
-  public Folder Folder;
+  public Folder Folder { get; set; }
 }
