+++ ApiDatastructures/Folders/Create/ResponseBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Folder.Create;
 
 public class ResponseBody
 {
-  public string FolderUuid;
+  public string FolderUuid { get; set; }
 }
