+++ ApiDatastructures/Folders/Get/RequestBody.cs
@@ -1,6 +1,6 @@
-namespace AnonKey_Backend.ApiDatastructures.Credentials.Get;
+namespace AnonKey_Backend.ApiDatastructures.Folder.Get;
 
 public class RequestBody
 {
-  public string FolderUuid;
+  public string FolderUuid { get; set; }
 }
