+++ ApiDatastructures/Folders/Get/ResponseBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Folder.Get;
 
 public class ResponseBody
 {
-  public Folder Folder;
+  public Folder Folder { get; set; }
 }
