+++ ApiDatastructures/Folders/Update/RequestBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Folder.Update;
 
 public class RequestBody
 {
-  public Folder Folder;
+  public Folder Folder { get; set; }
 }
