+++ ApiDatastructures/Folders/Update/ResponseBody.cs
@@ -2,5 +2,5 @@ namespace AnonKey_Backend.ApiDatastructures.Folder.Update;
 
 public class ResponseBody
 {
-  public string FolderUuid;
+  public string FolderUuid { get; set; }
 }
